Move button and custom color to own file
