+++ Native-Frameworks/Utilities/SystemImageName.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum SystemImageName: String {
+    case ellipsis = "ellipsis"
+    case xMark = "xmark"
+}
